# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_SignalPulseData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignalPulseData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[SignalPulseData]",
        "type": "object"
      },
      "ApiEnvelope_SignalPulseItem_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignalPulseItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[SignalPulseItem]",
        "type": "object"
      },
...
      "ApiEnvelope_StocksRadarData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StocksRadarData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[StocksRadarData]",
        "type": "object"
      },
...
      "ApiEnvelope_TokenRadarData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenRadarData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[TokenRadarData]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SignalPulseData": {
        "additionalProperties": true,
        "properties": {
          "has_more": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has More"
          },
          "health": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignalPulseHealth"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SignalPulseItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "query": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "returned_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned Count"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "title": "SignalPulseData",
        "type": "object"
      },
...
      "SignalPulseItem": {
        "additionalProperties": true,
        "properties": {
          "candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Id"
          },
          "candidate_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Score"
          },
          "candidate_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Type"
          },
          "created_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At Ms"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignalPulseDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Status"
          },
          "display_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Status"
          },
          "evidence_event_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Event Ids",
            "type": "array"
          },
          "evidence_packet_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Packet Hash"
          },
          "evidence_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Status"
          },
          "fact_card": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Card"
          },
          "factor_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factor Snapshot"
          },
          "gate": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate"
          },
          "gate_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Gate Reasons",
            "type": "array"
          },
          "gate_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Version"
          },
          "playbooks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Playbooks",
            "type": "array"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "pulse_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pulse Version"
          },
          "risk_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Risk Reasons",
            "type": "array"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "score_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Band"
          },
          "social_phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Social Phase"
          },
          "source_event_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Event Ids",
            "type": "array"
          },
          "subject": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "subject_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Key"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          },
          "updated_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At Ms"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "window": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window"
          }
        },
        "title": "SignalPulseItem",
        "type": "object"
      },
...
      "StocksRadarData": {
        "additionalProperties": true,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "projection": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "window": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window"
          }
        },
        "title": "StocksRadarData",
        "type": "object"
      },
...
      "TokenDiscussionDigestData": {
        "additionalProperties": true,
        "properties": {
          "coverage": {
            "additionalProperties": true,
            "title": "Coverage",
            "type": "object"
          },
          "currentness": {
            "$ref": "#/components/schemas/NarrativeCurrentnessData"
          },
          "data_gaps": {
            "items": {},
            "title": "Data Gaps",
            "type": "array"
          },
          "status": {
            "enum": [
              "ready",
              "pending",
              "insufficient",
              "semantic_unavailable",
              "stale"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "currentness"
        ],
        "title": "TokenDiscussionDigestData",
        "type": "object"
      },
      "TokenRadarData": {
        "additionalProperties": true,
        "properties": {
          "attention": {
            "items": {
              "$ref": "#/components/schemas/TokenRadarRowData"
            },
            "title": "Attention",
            "type": "array"
          },
          "projection": {
            "additionalProperties": true,
            "title": "Projection",
            "type": "object"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/TokenRadarRowData"
            },
            "title": "Targets",
            "type": "array"
          },
          "venue": {
            "title": "Venue",
            "type": "string"
          },
          "window": {
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "window",
          "scope",
          "venue"
        ],
        "title": "TokenRadarData",
        "type": "object"
      },
      "TokenRadarRowData": {
        "additionalProperties": true,
        "properties": {
          "discussion_digest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenDiscussionDigestData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "TokenRadarRowData",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Parallax",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/signal-lab/pulse": {
      "get": {
        "operationId": "signal_lab_pulse_api_signal_lab_pulse_get",
        "parameters": [
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "4h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "handle",
            "required": false,
            "schema": {
              "default": "",
              "title": "Handle",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "visibility",
            "required": false,
            "schema": {
              "default": "public",
              "title": "Visibility",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 80,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cursor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SignalPulseData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signal Lab Pulse",
        "tags": [
          "api"
        ]
      }
    },
    "/api/signal-lab/pulse/{candidate_id}": {
      "get": {
        "operationId": "signal_lab_pulse_by_id_api_signal_lab_pulse__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "visibility",
            "required": false,
            "schema": {
              "default": "public",
              "title": "Visibility",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SignalPulseItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signal Lab Pulse By Id",
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/stocks-radar": {
      "get": {
        "operationId": "stocks_radar_api_stocks_radar_get",
        "parameters": [
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "1h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_StocksRadarData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stocks Radar",
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/token-radar": {
      "get": {
        "operationId": "token_radar_api_token_radar_get",
        "parameters": [
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "1h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "venue",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Venue",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_TokenRadarData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Token Radar",
        "tags": [
          "api"
        ]
      }
    },
...
  }
}